+++ inst/htmlwidgets/lfx-wms/leaflet.wms.js
@@ -73,9 +73,14 @@ wms.Source = L.Layer.extend({
         overlayOptions.attribution = '';
         if (untiled) {
             return wms.overlay(this._url, overlayOptions);
-        } else {
-            return wms.tileLayer(this._url, overlayOptions);
         }
+        var tileOpts = L.extend({}, overlayOptions);
+        var hdrs = tileOpts.headers;
+        delete tileOpts.headers;
+        if (hasHeaders(hdrs)) {
+            return wms.tileLayerHeader(this._url, tileOpts, hdrs);
+        }
+        return wms.tileLayer(this._url, tileOpts);
     },
 
     'onAdd': function() {
@@ -175,9 +180,9 @@ wms.Source = L.Layer.extend({
         this.showWaiting();
         this.ajax(url, done);
 
-        function done(result) {
+        function done(result, status) {
             this.hideWaiting();
-            var text = this.parseFeatureInfo(result, url);
+            var text = this.parseFeatureInfo(result, url, status);
             callback.call(this, latlng, text);
         }
     },
@@ -215,11 +220,11 @@ wms.Source = L.Layer.extend({
         return L.extend({}, wmsParams, infoParams);
     },
 
-    'parseFeatureInfo': function(result, url) {
-        // Hook to handle parsing AJAX response
+    'parseFeatureInfo': function(result, url, status) {
         if (result == "error") {
-            // AJAX failed, possibly due to CORS issues.
-            // Try loading content in <iframe>.
+            if (status === 401 || status === 403) {
+                return "WMS GetFeatureInfo: unauthorized (" + status + ")";
+            }
             result = "<iframe src='" + upgradeInsecureUrl(url) + "' style='border:none'>";
         }
         return result;
@@ -313,6 +318,35 @@ wms.getSourceForUrl = function(url, options) {
 wms.TileLayer = L.TileLayer.WMS;
 wms.tileLayer = L.tileLayer.wms;
 
+wms.TileLayerHeader = L.TileLayer.WMS.extend({
+    initialize: function(url, options, headers) {
+        L.TileLayer.WMS.prototype.initialize.call(this, url, options);
+        this.headers = headers || [];
+    },
+    createTile: function(coords, done) {
+        var url = this.getTileUrl(coords);
+        var img = document.createElement('img');
+        img.setAttribute('role', 'presentation');
+        fetchBlob(url, this.headers, function(blob) {
+            if (!blob) {
+                done(new Error('WMS tile request failed'), img);
+                return;
+            }
+            var objUrl = URL.createObjectURL(blob);
+            img.onload = function() {
+                URL.revokeObjectURL(objUrl);
+            };
+            img.src = objUrl;
+            done(null, img);
+        });
+        return img;
+    }
+});
+
+wms.tileLayerHeader = function(url, options, headers) {
+    return new wms.TileLayerHeader(url, options, headers);
+};
+
 /*
  * wms.Overlay:
  * "Single Tile" WMS image overlay that updates with map changes.
@@ -337,7 +371,8 @@ wms.Overlay = L.Layer.extend({
         'opacity': 1,
         'isBack': false,
         'minZoom': 0,
-        'maxZoom': 18
+        'maxZoom': 18,
+        'headers': null
     },
 
     'initialize': function(url, options) {
@@ -408,34 +443,51 @@ wms.Overlay = L.Layer.extend({
             opt.zIndex=this.options.zIndex;
         if (this.options.pane)
             opt.pane=this.options.pane;
-        var overlay = L.imageOverlay(url, bounds, opt);
 
-        overlay.addTo(this._map);
-        overlay.once('load', _swap, this);
-        function _swap() {
-            if (!this._map) {
-                return;
-            }
-            if (overlay._url != this._currentUrl) {
-                this._map.removeLayer(overlay);
-                return;
-            } else if (this._currentOverlay) {
-                this._map.removeLayer(this._currentOverlay);
-            }
-            this._currentOverlay = overlay;
-            overlay.setOpacity(
-                this.options.opacity ? this.options.opacity : 1
-            );
-            if (this.options.isBack === true) {
-                overlay.bringToBack();
-            }
-            if (this.options.isBack === false) {
-                overlay.bringToFront();
+        var requestUrl = url;
+        var self = this;
+        function attachOverlay(src) {
+            var overlay = L.imageOverlay(src, bounds, opt);
+            overlay._wmsUrl = requestUrl;
+            overlay.addTo(self._map);
+            overlay.once('load', function() {
+                if (!self._map) {
+                    return;
+                }
+                if (overlay._wmsUrl != self._currentUrl) {
+                    self._map.removeLayer(overlay);
+                    if (src.indexOf('blob:') === 0) {
+                        URL.revokeObjectURL(src);
+                    }
+                    return;
+                } else if (self._currentOverlay) {
+                    self._map.removeLayer(self._currentOverlay);
+                }
+                self._currentOverlay = overlay;
+                overlay.setOpacity(
+                    self.options.opacity ? self.options.opacity : 1
+                );
+                if (self.options.isBack === true) {
+                    overlay.bringToBack();
+                }
+                if (self.options.isBack === false) {
+                    overlay.bringToFront();
+                }
+            });
+            if ((self._map.getZoom() < self.options.minZoom) ||
+                (self._map.getZoom() > self.options.maxZoom)){
+                self._map.removeLayer(overlay);
             }
         }
-        if ((this._map.getZoom() < this.options.minZoom) ||
-            (this._map.getZoom() > this.options.maxZoom)){
-            this._map.removeLayer(overlay);
+        if (hasHeaders(this.options.headers)) {
+            fetchBlob(url, this.options.headers, function(blob) {
+                if (!blob || requestUrl !== self._currentUrl) {
+                    return;
+                }
+                attachOverlay(URL.createObjectURL(blob));
+            });
+        } else {
+            attachOverlay(url);
         }
     },
 
@@ -500,6 +552,52 @@ wms.overlay = function(url, options) {
     return new wms.Overlay(url, options);
 };
 
+function headerList(headers) {
+    if (!headers) {
+        return [];
+    }
+    if (Object.prototype.toString.call(headers) === '[object Array]') {
+        return headers;
+    }
+    var out = [];
+    for (var key in headers) {
+        if (Object.prototype.hasOwnProperty.call(headers, key)) {
+            out.push({ header: key, value: headers[key] });
+        }
+    }
+    return out;
+}
+
+function hasHeaders(headers) {
+    return headerList(headers).length > 0;
+}
+
+function applyXhrHeaders(request, headers) {
+    headerList(headers).forEach(function(h) {
+        if (h && h.header && h.value != null) {
+            request.setRequestHeader(h.header, String(h.value));
+        }
+    });
+}
+
+function fetchBlob(url, headers, callback) {
+    var request = new XMLHttpRequest();
+    request.open('GET', url);
+    request.responseType = 'blob';
+    applyXhrHeaders(request, headers);
+    request.onreadystatechange = function() {
+        if (request.readyState !== 4) {
+            return;
+        }
+        if (request.status >= 200 && request.status < 300 && request.response) {
+            callback(request.response);
+        } else {
+            callback(null);
+        }
+    };
+    request.send();
+}
+
 function upgradeInsecureUrl(url) {
     if (typeof window !== 'undefined' &&
         window.location &&
@@ -537,6 +635,9 @@ function ajax(url, callback, redirectCount) {
         hops = redirectCount || 0;
     request.onreadystatechange = change;
     request.open('GET', url);
+    if (context && context.options) {
+        applyXhrHeaders(request, context.options.headers);
+    }
     request.send();
 
     function change() {
@@ -553,7 +654,7 @@ function ajax(url, callback, redirectCount) {
             if (loc) {
                 var next = absoluteUrl(url, loc);
                 if (!isSameHostRedirect(url, next)) {
-                    callback.call(context, 'error');
+                    callback.call(context, 'error', status);
                     return;
                 }
                 ajax.call(context, next, callback, hops + 1);
@@ -564,7 +665,7 @@ function ajax(url, callback, redirectCount) {
                 return;
             }
         }
-        callback.call(context, 'error');
+        callback.call(context, 'error', status);
     }
 }